fix: Support Unicode scripts well by switching to babel-minify
Utilise colourful output of babel-minify if parsing or mnification fails.

BREAKING CHANGE: Error messages of `terser` and `babel-minify` are very different. The new messages are more verbose and include a code exzerpt from the error occurrence. If you parsed the error messages, you will probably need to cut only the first line of the message, which contains the reason.