feat: Minify with swc by default
* Add swc as a JavaScript minifier
* Minify using swc by default

swc is the most efficient minifier. It proces the smallest size
and is only a little slower than esbuild. And it supports UMD
module format, which esbuild does not.

BREAKING CHANGE: Although the output of this tool did not change,
swc may have different behaviour than esbuild. The effect should
be better compression, but if this tool starts failing, you can
file a bugu about it and force esbuild or other minifier
as a temporary workaround.