feat: Use esbuild minified by default
BREAKING CHANGE: The default minifier is esbuild now. It will probably
break no real-world use case. Nevertheless, the reported minified
changes with the default parameters might change.

It does not need to specify if the source is a script or module.
It nmakes it as a default minifier more convenient.